services/auth/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.7.0
boto3>=1.34.0
